Externalize the internal stream class to another file
