+++ src/basics.js
@@ -2,89 +2,32 @@
 const { Readable } = require('stream');
 const { StreamError } = require('./errors');
 const { wrapInPromise } = require('./utils');
+const { InternalStream, STREAM_STATUS } = require('./internalStream');
 
+import type { Event, Producer as InternalStreamProducer, Push } from './internalStream';
 import type stream from 'stream';
 
-type EventDone = {| done: true |};
-type EventError  = {| error: Error, done?: false |};
-type EventValue<T>  = {| value: T, done?: false |};
-export type Event<T> = EventDone | EventError | EventValue<T>;
-
-const STREAM_STATUS = {
-  OPEN: 'OPEN',
-  DONE: 'DONE',
-  ERROR: 'ERROR'
-};
-
-type StreamStatus = $Keys<typeof STREAM_STATUS>;
-
-export type Push<T> = (Event<T>) => boolean;
-type InternalStreamProducer<ProducedT> = (push: Push<ProducedT>, context: { status: StreamStatus }) => void;
-
-class InternalStream<T> extends Readable {
-  status: StreamStatus;
-  constructor(producer: InternalStreamProducer<T> = (push) => {}) {
-    super({
-      objectMode: true,
-      read() {
-        producer(this.pushEvent.bind(this), this);
-      }
-    });
-    this.status = STREAM_STATUS.OPEN;
-  }
-  pushEvent(event: Event<T>): boolean {
-    if (this.status != STREAM_STATUS.OPEN) {
-      // console.log('InternalStream.pushEvent push when not open', event);
-      throw new StreamError('No event should be produced once the stream has ended.');
-    }
-    else if (event.done) {
-      // console.log('InternalStream.pushEvent done');
-      this.status = STREAM_STATUS.DONE;
-      this.push(null);
-      return false;
-    }
-    else if (event.error) {
-      // console.log('InternalStream.pushEvent', event.error.toString());
-      this.status = STREAM_STATUS.ERROR;
-      // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
-      this.push(event);
-      this.push(null);
-      return false;
-    }
-    else {
-      // console.log('InternalStream.pushEvent', event.value);
-      // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
-      return this.push(event);
-    }
-  }
-}
-
-export type ReducerResult<T> = Promise<{| accumulation: T, done: ?boolean |}> | {| accumulation: T, done: ?boolean |}
-export type Reducer<T, AccumulationT> = (AccumulationT, Event<T>) => ReducerResult<AccumulationT>;
-
-function reduceInternalStream<T>(stream: InternalStream<T>, event: Event<T>): ReducerResult<InternalStream<T>> {
-  // Only push when the stream is open !
-  if (stream.status != STREAM_STATUS.OPEN) {
-    return { accumulation: stream, done: true };
-  }
-  else {
-    stream.pushEvent(event);
-    const done = event.done || (event.error && true);
-    return { accumulation: stream, done };
-  }
-}
-
-opaque type StreamInternals<T> = {
-  stream: Promise<stream.Readable>,
-  consumer?: any
-}
-
 export interface Stream<T> {
-  internals: StreamInternals<T>,
+  stream: Promise<InternalStream<T>>,
+  consumer?: any,
   countConsumers(): number,
   thru<R, Fn: (Stream<T>) => R>(f: Fn): R
 }
 
+function createStream<T>(stream: InternalStream<T> | Promise<InternalStream<T>>): Stream<T> {
+  return {
+    stream: stream instanceof Promise ? stream : Promise.resolve(stream),
+    countConsumers() {
+      return this.internals.consumer ? 1 : 0;
+    },
+    thru<R, Fn: (Stream<T>) => R>(f: Fn): R {
+      return f(this);
+    }
+  };
+}
+
+export type ReducerResult<T> = Promise<{| accumulation: T, done: ?boolean |}> | {| accumulation: T, done: ?boolean |}
+export type Reducer<T, AccumulationT> = (AccumulationT, Event<T>) => ReducerResult<AccumulationT>;
 export type ReducerTransformer<T, TransformedT, AccumulationT> = (Reducer<TransformedT, AccumulationT>) => Reducer<T, AccumulationT>;
 export type Seeder<AccumulationT> = () => AccumulationT;
 
@@ -178,18 +121,30 @@ function transduce<T, TransformedT, AccumulationT>(
   });
 
   return (stream) => {
-    if (stream.internals.consumer) {
+    if (stream.consumer) {
       throw new StreamError('Stream already being consumed.');
     }
-    const transducerPromise = stream.internals.stream.then(consumeStream);
-    stream.internals.consumer = transducerPromise;
+    const transducerPromise = stream.stream.then(consumeStream);
+    stream.consumer = transducerPromise;
     return transducerPromise.then(({ accumulation }) => accumulation);
   };
 }
 
+function reduceInternalStream<T>(stream: InternalStream<T>, event: Event<T>): ReducerResult<InternalStream<T>> {
+  // Only push when the stream is open !
+  if (stream.status != STREAM_STATUS.OPEN) {
+    return { accumulation: stream, done: true };
+  }
+  else {
+    stream.pushEvent(event);
+    const done = event.done || (event.error && true);
+    return { accumulation: stream, done };
+  }
+}
+
 function transduceToStream<ConsumedT, ProducedT>(transformer: ReducerTransformer<ConsumedT, ProducedT, any>): (Stream<ConsumedT>) => Stream<ProducedT> {
   const transducer = transduce(transformer, reduceInternalStream, () => new InternalStream());
-  return (stream) => wrapReadableStream(transducer(stream));
+  return (stream) => createStream(transducer(stream));
 }
 
 export type Transformer<ConsumedT, ProducedT, SeedT = void> = (event: Event<ConsumedT>, push: Push<ProducedT>, seed: ?SeedT) => ?SeedT | Promise<?SeedT>;
@@ -242,20 +197,6 @@ function subscribe<T>(subscriber: Subscriber<T>): (Stream<T>) => Promise<void> {
     () => undefined);
 }
 
-function wrapReadableStream<T>(stream: stream.Readable | Promise<stream.Readable>): Stream<T> {
-  return {
-    internals: {
-      stream: stream instanceof Promise ? stream : Promise.resolve(stream)
-    },
-    countConsumers() {
-      return this.internals.consumer ? 1 : 0;
-    },
-    thru<R, Fn: (Stream<T>) => R>(f: Fn): R {
-      return f(this);
-    }
-  };
-}
-
 export type Producer<ProducedT, SeedT> = (push: Push<ProducedT>, seed: ?SeedT) => ?SeedT | Promise<?SeedT>;
 
 function produce<ProducedT, SeedT>(producer: Producer<ProducedT, SeedT>, seed: ?SeedT): Stream<ProducedT> {
@@ -281,7 +222,7 @@ function produce<ProducedT, SeedT>(producer: Producer<ProducedT, SeedT>, seed: ?
     }
   };
 
-  return wrapReadableStream(new InternalStream(internalProducer));
+  return createStream(new InternalStream(internalProducer));
 }
 
 function fromReadable<T>(readable: stream.Readable): Stream<T> {
@@ -309,23 +250,23 @@ function fromReadable<T>(readable: stream.Readable): Stream<T> {
     .on('end', endListener)
     .resume();
 
-  return wrapReadableStream(internalStream);
+  return createStream(internalStream);
 }
 
 function fromIterable<T>(iterable: Iterable<T>): Stream<T> {
   // $FlowFixMe bug in the Iterable type of flow (cf. https://github.com/facebook/flow/issues/1163)
   const iterator: Iterator<T> = iterable[Symbol.iterator]();
-  return wrapReadableStream(new InternalStream(
-    (push) => {
+  return createStream(new InternalStream(
+    (push: Push<T>) => {
       let continueProduction = true;
       while (continueProduction) {
-        const { done, value } = iterator.next();
-        if (done) {
+        const iteratorResult = iterator.next();
+        if (iteratorResult.done) {
           push({ done: true });
           continueProduction = false;
         }
         else {
-          continueProduction = push({ value });
+          continueProduction = push({ value: iteratorResult.value });
         }
       }
     }
